+++ scripts/ssh_key_manager.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # Set base directories
-BASE_DIR=/opt/apps/authentipy
+BASE_DIR=/opt/apps/insrt
 VENV_DIR=$BASE_DIR/venv
 LOGS_DIR=/opt/apps/resources/logs
 SCRIPT_DIR=$BASE_DIR/scripts